Revert "CAP-79 #comment Organize files"
This reverts commit 8d8c8a11
